React Apps have hot relaod
Hot reload of React Apps in the react folder, hot reload of css, ability to run in production mode and dev mode. Production mode turns off hot reload capabilities, updated the package.json scripts for building/starting and updated the webpack configs to use dotenv configs for building. Finally, modified my server.js so that it would use webpack-dev-middleware and webpack-hot-middleware to compile the target of webpack.config.js for hot reloading
