+++ .babelrc
@@ -1,4 +1,4 @@
 {
   "presets": ["@babel/preset-env", "@babel/preset-react"],
-  "plugins": ["@babel/plugin-transform-runtime", "@babel/plugin-transform-react-jsx"]
+  "plugins": ['react-hot-loader/babel', "@babel/plugin-transform-runtime", "@babel/plugin-transform-react-jsx"]
 }
