+++ package.json
@@ -12,14 +12,10 @@
 	},
 	"scripts": {
 		"clean": "rimraf bin/",
-		"webpack-build": "webpack --config webpack.config.js",
-		"webpack-build-server": "webpack --config webpack.config.server.js",
-		"babel-build": "babel ./src/server --out-dir bin/server --source-maps",
-		"dev-build": "rimraf bin/ && babel ./src --out-dir bin/ --copy-files --source-maps",
-		"build": "npm run clean && npm run webpack-build && npm run babel-build",
-		"build-and-watch": "npm run clean && npm run webpack-build --watch & npm run babel-build --watch",
-		"start": "nodemon  -r dotenv/config --inspect bin/server",
-		"start-webpack": "node -r dotenv/config --inspect bin/server/server",
+		"build-client": "webpack  --config  webpack.config.js",
+		"build-server": "webpack  --config webpack.config.server.js",
+		"clean-build": "run-s clean  build-client build-server",
+		"start": "node -r dotenv/config --inspect bin/server/server",
 		"test": "jest --detectOpenHandles",
 		"test-url": "jest shorturl-dao.test.js"
 	},
@@ -47,8 +43,11 @@
 		"mongoose": "^5.9.3",
 		"morgan": "~1.9.1",
 		"node-sass": "^4.13.1",
+		"npm-run-all": "^4.1.5",
 		"react": "^16.12.0",
 		"react-dom": "^16.12.0",
+		"react-hot-loader": "^4.12.20",
+		"@hot-loader/react-dom": "^16.8.6",
 		"react-redux": "^7.2.0",
 		"react-scripts": "3.3.1",
 		"redux": "^4.0.5",
@@ -85,7 +84,11 @@
 		"jest-babel": "^1.0.1",
 		"nodemon": "^2.0.2",
 		"rimraf": "^2.6.3",
+		"style-loader": "^1.1.3",
 		"webpack": "^5.0.0-beta.14",
-		"webpack-cli": "^3.3.11"
+		"webpack-cli": "^3.3.11",
+		"webpack-dev-middleware": "^3.7.2",
+		"webpack-hot-middleware": "^2.25.0",
+		"webpack-node-externals": "^1.7.2"
 	}
 }
