+++ src/app/calculator/style.scss
@@ -32,10 +32,10 @@ html {
 }
 
 .themeFour {
-  --shade1: #e6e6e6;
-  --shade2: #dedede;
-  --shade3: #d1d1d1;
-  --shade4: #c7c7c7;
+  --shade1: #466b8c;
+  --shade2: #8c5f46;
+  --shade3: #8c8346;
+  --shade4: #79468c;
    --grey1: #dedede;
 }
