+++ webpack.config.server.js
@@ -1,17 +1,40 @@
+const dotenv = require('dotenv');
+const result = dotenv.config();
+if (result.error) {
+  throw result.error;
+}
+const { parsed: envs } = result;
+
+console.log("webpack.config.js envs are " + JSON.stringify(envs))
+
+var nodeEnv = envs.NODE_ENV;
+const isProduction = nodeEnv !== 'development';
+if(isProduction)
+  nodeEnv = 'production';
+
+console.log("setting up env for : " + nodeEnv);
+
 
 const entryObject = {}
 const path = require('path');
+const webpack = require('webpack');
+
+var nodeExternals = require("webpack-node-externals");
 
 entryObject.index  = {
   import: "./src/server/server.js",
   filename: 'server/server.js'
 };
 
-console.log("entryObject  is " + JSON.stringify(entryObject));
+console.log("Setting up env for server build in  " + nodeEnv);
+const plugins = [];
+
+if (!isProduction) {
+    plugins.push(new webpack.HotModuleReplacementPlugin())
+}
 
 module.exports = {
   entry: entryObject,
-  target: "node",
   output: {
     filename: '[name]',
     path: path.resolve(__dirname, 'bin')
@@ -27,9 +50,18 @@ module.exports = {
       }
     ]
   },
+  watch: false,
+  watchOptions: {
+  	ignored: ['node_modules/*']
+  },
+  target: "node",
+  externals: [nodeExternals()],
+  plugins:plugins,
+  node: {
+  	__dirname: false,
+  	__filename: false,
+    global: true
 
-  plugins: [
-
-  ],
-  mode: "development"
+  },
+  mode: nodeEnv
 };
